fix(gemm): q4 coalesced gemv scales are f16, not f32 (revert #33) (#36)
#33 changed mt_gemv_q4_coalesced/_relu2/_accum to read scales as
Tensor<f32>, on the evidence that the param is named `d_f32` and the
in-file perf benches bind f32. But the production consumer — the resident
weight decode/prefill in FFAI's Nemotron model — uploads these scales as
**f16** (half the bytes), and the perf benches don't check correctness so
their f32 binding was a latent no-op. Reading f16 scale bytes as f32
produced garbage scales → NaN from the first Mamba in_proj gemv → NaN
prefill logits and a corrupted decode argmax (131071 instead of the HF
1234).

Revert the three kernels to f16 and document why the name is historical.
Verified on a 5090: Nemotron decode argmax returns to 1234 (= HF) and the
prefill correctness gate passes (cosine 0.994, seq==batched argmax).

The synthetic gemv_q4 correctness test that fed f32 (and motivated #33)
is corrected to feed f16 in the matching FFAI PR.

+++ crates/ffai-kernels-std/src/kernels/gemm/gemv_quantized.rs
@@ -238,7 +238,10 @@ pub fn mt_gemv_q8_coalesced_accum<T>(
 #[kernel]
 pub fn mt_gemv_q4_coalesced<T>(
     qs: Tensor<u32>,
-    d_f32: Tensor<f32>,
+    // f16 scales (half the bytes — the resident-weight decode/prefill path uploads
+    // them as f16). Param name is historical; do NOT change to Tensor<f32> — the
+    // production callers feed f16 and reading f16 bytes as f32 yields NaN.
+    d_f32: Tensor<f16>,
     x: Tensor<T>,
     mut out: Tensor<T>,
     #[constexpr] k_in: u32,
@@ -432,7 +435,10 @@ pub fn mt_gemv_q4_coalesced_2row<T>(
 #[kernel]
 pub fn mt_gemv_q4_coalesced_relu2<T>(
     qs: Tensor<u32>,
-    d_f32: Tensor<f32>,
+    // f16 scales (half the bytes — the resident-weight decode/prefill path uploads
+    // them as f16). Param name is historical; do NOT change to Tensor<f32> — the
+    // production callers feed f16 and reading f16 bytes as f32 yields NaN.
+    d_f32: Tensor<f16>,
     x: Tensor<T>,
     mut out: Tensor<T>,
     #[constexpr] k_in: u32,
@@ -482,7 +488,10 @@ pub fn mt_gemv_q4_coalesced_relu2<T>(
 #[allow(clippy::too_many_arguments)]
 pub fn mt_gemv_q4_coalesced_accum<T>(
     qs: Tensor<u32>,
-    d_f32: Tensor<f32>,
+    // f16 scales (half the bytes — the resident-weight decode/prefill path uploads
+    // them as f16). Param name is historical; do NOT change to Tensor<f32> — the
+    // production callers feed f16 and reading f16 bytes as f32 yields NaN.
+    d_f32: Tensor<f16>,
     x: Tensor<T>,
     mut acc: Tensor<T>,
     scale: Tensor<f32>,
